requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pandas==2.2.0
cloudpickle==3.0.0
supabase==2.3.0
python-dotenv==1.0.0
scikit-learn==1.4.0
python-multipart==0.0.6
xgboost==2.0.3
